+++ src/app/permisos/index.jsx
@@ -42,3 +42,4 @@ const PermissionsSystem = () => {
 }
 
 export default PermissionsSystem
+
